+++ scripts/gmactions.js
@@ -1,4 +1,4 @@
-import { setting } from '../innocenti-loot.js';
+import { setting, moduleName, LOOTED } from '../innocenti-loot.js';
 export class GMActions {
     constructor(data = {}) {
         this.data = data;
@@ -8,17 +8,22 @@ export class GMActions {
             this.token = canvas.tokens.get(data.tokenid);
             this.actor = game.actors.contents.find(a => a.id === this.token.actor.id);
         }
+        if (data.targets) {
+            this.targets = data.targets;
+        }
     }
 
     async SendLoot() {
         if (setting('debug')) console.log("DATA", this.data)
+        await this.FlagTargets()
         let actor = game.actors.get(this.data.actor);
         await actor.update({ "system.currency": this.data.currency });
         await actor.createEmbeddedDocuments("Item", this.data.items, { noHook: true });
     }
 
     async CreateLoot() {
-        if (setting('debug')) console.log("Create - DATA", this.data)
+        if (setting('debug')) console.log("Create - DATA", this.data);
+        await this.FlagTargets();
         let actors = await Actor.create({
             name: this.data.lootName,
             type: "npc",
@@ -28,9 +33,9 @@ export class GMActions {
         });
         let lootingUsers = game.users.contents.filter(user => { return user.role >= 1 && user.role <= 2 });
         let permissions = {};
-        Object.assign(permissions, actors.data.permission);
+        Object.assign(permissions, actors._source.permission);
         lootingUsers.forEach(user => {
-            permissions[user.id] = CONST.ENTITY_PERMISSIONS.OBSERVER;
+            permissions[user._id] = CONST.DOCUMENT_PERMISSION_LEVELS.OBSERVER;
         });
         await actors.update({ "system.currency": this.data.currency, permission: permissions });
         await actors.createEmbeddedDocuments("Item", this.data.items, { noHook: true });
@@ -38,5 +43,12 @@ export class GMActions {
         if (!setting('debug')) canvas.tokens.deleteMany(this.data.targets);
 
         if (setting('debug')) console.log("Fim da chamada", this.data);
+    }
+
+    async FlagTargets() {
+        for (var i = 0; i < this.targets.length; i++) {
+            let target = canvas.tokens.get(this.targets[i].id);
+            target.document.setFlag(moduleName, LOOTED, true);
+        }
     }
 }